KAFKA-4596: Cleaned up logic
