+++ core/src/main/scala/kafka/admin/ReassignPartitionsCommand.scala
@@ -359,20 +359,22 @@ class ReassignPartitionsCommand(zkUtils: ZkUtils, proposedAssignment: Map[TopicA
     }
   }
 
+
   private def postRebalanceReplicasThatMoved(existing: Map[TopicAndPartition, Seq[Int]], proposed: Map[TopicAndPartition, Seq[Int]]): Map[TopicAndPartition, Seq[Int]] = {
     //For each partition in the proposed list, filter out any replicas that exist now (i.e. are in the proposed list and hence are not moving)
-    existing
-      .filter{ case (tp, current) => proposed.contains(tp)}
-      .map { case (tp, current) =>
-      tp -> (proposed(tp).toSet -- current).toSeq
+    proposed.map { case (tp, proposedReplicas) =>
+      tp -> (proposedReplicas.toSet -- existing(tp)).toSeq
     }
   }
 
   private def preRebalanceReplicaForMovingPartitions(existing: Map[TopicAndPartition, Seq[Int]], proposed: Map[TopicAndPartition, Seq[Int]]): Map[TopicAndPartition, Seq[Int]] = {
-    //Throttle all existing replicas (as any one might be a leader). So just filter out those which aren't moving
-    existing.filter { case (tp, current) =>
-      proposed.contains(tp) &&
-      (proposed(tp).toSet -- current).nonEmpty
+    //For any moving partition, throttle all the original (pre move) replicas (as any one might be a leader)
+
+    def moving(postMoveReplicas: Seq[Int], preMoveReplicas: Seq[Int]) =
+      (postMoveReplicas.toSet -- preMoveReplicas.toSet).nonEmpty
+
+    existing.filter { case (tp, preMoveReplicas) =>
+      proposed.contains(tp) && moving(preMoveReplicas, proposed(tp))
     }
   }
